moved into skill.c
