+++ server/src/Makefile
@@ -10,7 +10,7 @@ C_FLAGS = $(O_FLAGS) -pedantic-errors -Wall \
 L_FLAGS = $(LINK) $(DEBUG) $(PROF) -lcrypt -lm
 
 O_FILES = act_comm.o act_info.o act_move.o act_obj.o act_wiz.o board.o comm.o const.o \
-          create.o healer.o hunt.o db.o fight.o handler.o interp.o magic.o mob_commands.o \
+          healer.o hunt.o db.o fight.o handler.o interp.o magic.o mob_commands.o \
           mob_prog.o save.o special.o update.o quest.o trap.o skill.o sft.o table.o \
           deity.o tournament.o bot.o
